Multiple improvements in zoning, chair use and subskill increase
git-svn-id: http://www.linux-addicted.net/repos/tinns@28 ef8e7eb3-8d0b-0410-9eab-9562649fb893

+++ trunk/tinns-v2/src/game/decoder/udp_subskill.h
@@ -36,7 +36,7 @@
 class PUdpSubskillInc : public PUdpMsgAnalyser
 {
   private:
-    u8 SubskillID;
+    u16 SubskillID;
     
   public:
     PUdpSubskillInc(PMsgDecodeData* nDecodeData);
